refactor(web): extract logical-volume-form transformations
Moves data transformation logic from Form.tsx to a new transformations.ts
module:

- buildPayload: converts form values to ConfigModel.LogicalVolume
- toFormValues: converts ConfigModel.LogicalVolume to initial form values
- useSolvedSizes: calculates min/max sizes for logical volumes
- lvNameFromMountPoint: derives LV name from mount point (auto-fill helper)

Like partition-form, this module uses shared helpers for filesystem and size
configuration, adding LV-specific logic for target and lvName fields.

Also updates Form.test.tsx to mock the new transformations module instead of
the old use-solved-sizes module.

Reduces Form.tsx from 482 to ~360 lines.
